Start refactoring into smaller components
