+++ src/shared/components/SearchComponents/FindMyLocation.tsx
@@ -1,0 +1,24 @@
+import { Box, IconButton, InputBase, OutlinedInput } from '@mui/material';
+import React from 'react';
+import { useDeviceSizes } from 'shared/hooks';
+import MyLocationIcon from '@mui/icons-material/MyLocation';
+
+export const FindMyLocation: React.FC = () => {
+  const { mediumDeviceSize } = useDeviceSizes();
+
+  return (
+    <Box sx={{ display: 'flex', pl: 3, pr: 1.5 }}>
+      {!mediumDeviceSize && (
+        <>
+          <IconButton disabled>
+            <MyLocationIcon />
+          </IconButton>
+          <InputBase placeholder="Find my location" />
+        </>
+      )}
+      {mediumDeviceSize && (
+        <OutlinedInput fullWidth placeholder="Find my location" />
+      )}
+    </Box>
+  );
+};
