+++ src/shared/components/index.ts
@@ -1,6 +1,5 @@
 export * from './Copyright';
 export * from './CustomAlert';
 export * from './CustomSnackbar';
-export * from './LandingHero';
 export * from './NavComponents';
-export * from './SearchFacility';
+export * from './SearchComponents';
